requirements.txt
alembic==1.5.8
anyio==3.6.1
attrs==20.3.0
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet>=3.0.2,<5
click==7.1.2
dnspython==2.1.0
docstring-parser==0.7.3
email-validator==1.1.2
Faker==7.0.0
fastapi==0.85.0
FastAPI-SQLAlchemy==0.2.1
greenlet==1.0.0
h11>=0.8,<0.10
idna==2.10
importlib-metadata==5.0.0
iniconfig==1.1.1
sentence_transformers==3.0.0
Mako==1.1.4
MarkupSafe==1.1.1
numpy==1.24.4
packaging
pandas==1.3.5
passlib==1.7.4
pluggy==0.13.1
psycopg2-binary==2.9.9
py==1.10.0
pycparser==2.20
pydantic==1.10.2
PyJWT==2.0.1
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-multipart==0.0.5
pytz==2022.4
requests==2.25.1
six==1.15.0
sniffio==1.3.0
SQLAlchemy<2.0
starlette==0.20.4
text-unidecode==1.3
toml==0.10.2
typing_extensions
urllib3==1.26.4
uvicorn==0.13.4
zipp==3.9.0
pytesseract~=0.3.10
pdf2image~=1.17.0
nltk~=3.8.1
elasticsearch~=8.13.1
python-docx
huggingface_hub==0.23.2
pyvi~=0.1.1
PyPDF2
oauth2client~=4.1.3
pydrive~=1.3.1
docx~=0.2.4
cryptography>=36.0.0
google-api-python-client~=2.131.0
openai==0.28.0
tenacity~=8.3.0
transformers~=4.41.1
google-cloud-vision==3.7.2
grpcio==1.64.1
grpcio-status==1.62.2
asyncpg==0.29.0
APScheduler==3.10.4
reportlab==4.2.2
pdfplumber==0.11.2
python-pptx==1.0.0
pyth==0.6.0
openpyxl==3.1.5
googletrans==4.0.0-rc1
httpx==0.13.3
scikit-learn~=1.3.2
scikit-image~=0.21.0
opencv-python~=4.10.0.84